fixes for cuda 12.0 on ubuntu 20.04 and boost 1.71 and higher compute capable systems

+++ packages/gpu_voxels/src/gpu_visualization/Visualizer.cu
@@ -1684,7 +1684,7 @@ void Visualizer::renderFunction(void)
   {
     float wait = 1000.f / m_max_fps - m_delta_time;
     m_delta_time += wait;
-    boost::this_thread::sleep(boost::posix_time::milliseconds(wait));
+    boost::this_thread::sleep(boost::posix_time::milliseconds(long(wait)));
   }
 }
